+++ src/app/content/components/Toolbar.spec.tsx
@@ -12,8 +12,6 @@ import { makeSearchResults } from '../../../test/searchResults';
 import TestContainer from '../../../test/TestContainer';
 import * as Services from '../../context/Services';
 import { receiveFeatureFlags } from '../../featureFlags/actions';
-import * as featureFlagSelectors from '../../featureFlags/selectors';
-import { searchButtonColor } from '../../featureFlags/selectors';
 import { MiddlewareAPI, Store } from '../../types';
 import { assertDocument, assertWindow } from '../../utils';
 import { practiceQuestionsFeatureFlag } from '../constants';
@@ -319,8 +317,8 @@ describe('search button', () => {
   const render = () => renderer.create(<TestContainer store={store}><Toolbar /></TestContainer>);
 
   it('button has theme bg color applied', () => {
-    const color = searchButtonColor.resultFunc('bannerColorButton', book, 'blue');
-    jest.spyOn(featureFlagSelectors, 'searchButtonColor').mockReturnValue(color);
+    const color = searchSelectors.searchButtonColor.resultFunc('bannerColorButton', book, 'blue');
+    jest.spyOn(searchSelectors, 'searchButtonColor').mockReturnValue(color);
     jest.spyOn(searchSelectors, 'mobileToolbarOpen').mockReturnValue(true);
 
     const component = render();
@@ -331,8 +329,8 @@ describe('search button', () => {
   });
 
   it('button has gray bg color applied', () => {
-    const color = searchButtonColor.resultFunc('grayButton', book, 'red');
-    jest.spyOn(featureFlagSelectors, 'searchButtonColor').mockReturnValue(color);
+    const color = searchSelectors.searchButtonColor.resultFunc('grayButton', book, 'red');
+    jest.spyOn(searchSelectors, 'searchButtonColor').mockReturnValue(color);
 
     const component = render();
     const [searchButton, searchButtonMobile] = component.root.findAllByType(SearchButton);
@@ -342,8 +340,8 @@ describe('search button', () => {
   });
 
   it('button has no bg color applied', () => {
-    const color = searchButtonColor.resultFunc(null, book, 'blue');
-    jest.spyOn(featureFlagSelectors, 'searchButtonColor').mockReturnValue(color);
+    const color = searchSelectors.searchButtonColor.resultFunc(null, book, 'blue');
+    jest.spyOn(searchSelectors, 'searchButtonColor').mockReturnValue(color);
 
     const component = render();
     const [searchButton, searchButtonMobile] = component.root.findAllByType(SearchButton);
@@ -353,7 +351,7 @@ describe('search button', () => {
   });
 
   it('new search buttons use new close button', () => {
-    jest.spyOn(featureFlagSelectors, 'searchButtonColor').mockReturnValue('green');
+    jest.spyOn(searchSelectors, 'searchButtonColor').mockReturnValue('green');
 
     const component = render();
     const findById = makeFindByTestId(component.root);
