+++ src/app/content/studyGuides/reducer.spec.ts
@@ -1,6 +1,4 @@
-import { receiveFeatureFlags } from '../../featureFlags/actions';
 import { locationChange } from '../../navigation/actions';
-import { studyGuidesFeatureFlag } from '../constants';
 import { HighlightData, SummaryHighlights } from '../highlights/types';
 import * as actions from './actions';
 import reducer, { initialState } from './reducer';
@@ -27,11 +25,6 @@ describe('study guides reducer', () => {
     const state = reducer(undefined, actions.receiveSummaryStudyGuides(summary, {pagination: null}));
 
     expect(state.summary.studyGuides).toEqual(summary);
-  });
-
-  it('enables study guides on receiveFeatureFlags', () => {
-    const state = reducer(initialState, receiveFeatureFlags([studyGuidesFeatureFlag]));
-    expect(state.isEnabled).toEqual(true);
   });
 
   it('sets summary filters', () => {
