+++ src/app/content/studyGuides/selectors.ts
@@ -1,4 +1,6 @@
 import { createSelector } from 'reselect';
+import * as selectFeatureFlags from '../../featureFlags/selectors';
+import { studyGuidesFeatureFlag } from '../constants';
 import { getHighlightColorFiltersWithContent, getHighlightLocationFilterForPage } from '../highlights/utils';
 import {
   getHighlightLocationFilters,
@@ -20,8 +22,8 @@ export const localState = createSelector(
 );
 
 export const studyGuidesEnabled = createSelector(
-  localState,
-  (state) => state.isEnabled
+  selectFeatureFlags.localState,
+  (state) => !!state[studyGuidesFeatureFlag]
 );
 
 export const studyGuidesSummary = createSelector(
