+++ src/app/content/studyGuides/types.ts
@@ -12,7 +12,6 @@ export interface StudyGuidesSummaryFilters {
 }
 
 export interface State {
-  isEnabled: boolean;
   summary: {
     filters: StudyGuidesSummaryFilters,
     loading: boolean,
