+++ src/app/featureFlags/selectors.ts
@@ -1,31 +1,19 @@
 import { createSelector } from 'reselect';
-import { book, bookTheme } from '../content/selectors';
-import { BookWithOSWebData } from '../content/types';
 import * as parentSelectors from '../selectors';
 
-export const enabled = createSelector(
+export const localState = createSelector(
   parentSelectors.localState,
   (parentState) => parentState.featureFlags
 );
 
 export const searchButtonStyle = createSelector(
-  enabled,
+  localState,
   (featureFlags) => featureFlags.searchButton || null
 );
 
 export const kineticBannerVariant = createSelector(
-  enabled,
+  localState,
   (featureFlags) => featureFlags.kineticEnabled && typeof featureFlags.kineticBanner === 'number'
     ? featureFlags.kineticBanner
     : false
-);
-
-export const searchButtonColor = createSelector(
-  searchButtonStyle,
-  book,
-  bookTheme,
-  (selectedStyle, selectedBook, selectedTheme) =>
-    selectedBook && selectedStyle === 'grayButton' ? 'gray' as BookWithOSWebData['theme']
-      : (selectedBook && selectedStyle === 'bannerColorButton' ? selectedTheme : null
-  )
 );
